Use full path on redirect
The document ID is the important part of the /doc/:id-ish route. This is
what we use to find the Outline document. But in the navigation, like in
the Outline UI, we prefix the ID with a kebabified version of the page
title to make the content of the page a bit more transparent in the URL

In the URL mapper, we were stripping out this detail, redirecting just
to /doc/:id. This change brings redirects into line with the approach to
the navigation URLs