Always show current page in navigation
The page navigation can get pretty long when you go a level or two deep
and/or there are lots of sibling or child pages. Currently, when you get
lower down the tree, sometimes the current page isn't visible in the
navigation when the page loads. This ensures we scroll it into view when
JavaScript is enabled. In practice and at least on my device, the scroll
doesn't seem to be perceptible, so it's as though the page loaded in the
correct place